Fix integer literals being only 32-bit.
Also changed Reservation::emitPutValue to not use Reservation pointers.

+++ src/CompileState.hpp
@@ -120,10 +120,10 @@ public:
 class LiteralNode {
 public:
     LiteralType type;
-    int i;
+    long i;
     char c;
 
-    LiteralNode(int i);
+    LiteralNode(long i);
     LiteralNode(char c);
 };
